requirements.txt
bcrypt==3.2.0
cffi==1.15.0
cryptography==36.0.2
fabric==2.7.0
invoke==1.7.0
paramiko==2.10.3
pathlib2==2.3.7.post1
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==0.20.0
six==1.16.0
